pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comic-baba"
version = "0.1.0"
description = "Temporal Hallucinations: character-consistent frame interpolation for comic clips"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "numpy>=1.24",
    "Pillow>=10.0",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
video = [
    "opencv-python-headless>=4.8",
]
dev = [
    "pytest>=7.4",
    "ruff>=0.4",
]

[project.scripts]
comic-baba = "comic_baba.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
